[Refactor] Rename Graphics classes to *Resource (Mesh, Texture, Material)
- Rename Mesh, Texture, and Material classes in the Graphics module to MeshResource, TextureResource, and MaterialResource

- Update all corresponding filenames, include paths, and project references (.vcxproj, .filters)

+++ Engine/src/Graphics/DebugDraw/DebugRenderer.cpp
@@ -270,7 +270,7 @@ namespace Graphics
 		}
 
 		// 프레임 상태 초기화
-		mCurrentFrameIndex = mDevice->GetSwapChain()->GetCurrentBackBufferIndex();
+		mCurrentFrameIndex = static_cast<uint32>(mDevice->GetSwapChain()->GetCurrentBackBufferIndex());
 		mCurrentDrawIndex = 0;
 
 		// PSO 및 공통 상태 설정
